7_calculation

+++ app/src/main/java/com/egorpoprotskiy/eyeofwages/data/Month.kt
@@ -1,12 +1,12 @@
 package com.egorpoprotskiy.eyeofwages.data
 
 data class Month (
-    val oklad: Int,
+    val oklad: Double,
     val norma: Int,
     val rabTime: Int,
     val nochTime: Int,
     val prazdTime: Int,
-    val premia: Int,
+    val premia: Double,
     val visluga: Int,
     val prikazDen: Int,
     val prikazNoch: Int
